# backend/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Your Pet",
    "description": "Your Pet project"
  },
  "consumes": ["application/json", "multipart/form-data"],
  "produces": ["application/json"],
  "servers": [
    { "url": "https://mypet.up.railway.app" },
    { "url": "http://localhost:3000" }
  ],
...
  "paths": {
...
    "/notices/{noticeId}": {
      "get": {
        "tags": ["Notice"],
        "description": "Get one notice by id (Public)",
        "parameters": [{ "$ref": "#/components/parameters/noticeId" }],
        "responses": {
          "200": {
            "description": "Success, received one notice",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notice": {
                      "$ref": "#/components/responses/NoticeExtendedResponse"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "404": {
            "description": "No notice found for this id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notice not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "delete": {
        "tags": ["Notice"],
        "security": [{ "Bearer": [] }],
        "description": "Delete notice (Private)",
        "parameters": [{ "$ref": "#/components/parameters/noticeId" }],
        "responses": {
          "200": {
            "description": "Success, deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "notice": { "$ref": "#/components/responses/UserResponse" }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/NotAuthorized"
          },
          "403": {
            "description": "The person who deletes the publication is not its owner",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Notice owner is not you"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Notice not found (so cannot be deleted)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "parameters": {
      "noticeId": {
        "name": "noticeId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "example": "64bccc6f1ab4316641be2221"
        }
      },
...
    },
    "responses": {
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": "64baea04ba92a3f0060ba5b7"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 16,
            "example": "John"
          },
          "email": {
            "type": "string",
            "format": "email",
            "pattern": "^\\w+([\\.-]?\\w+)*@\\w+([\\.-]?\\w+)*(\\.\\w{2,3})+$",
            "example": "johnnattas1987@gmail.com"
          },
          "city": {
            "type": "string",
            "minLength": 2,
            "pattern": "^[a-zA-Z]+$",
            "example": "Kyiv"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+\\d{12}$",
            "example": "+380984739205"
          },
          "birthday": {
            "type": "string",
            "pattern": "^\\d{2}-\\d{2}-\\d{4}$",
            "example": "18-08-1987"
          },
          "avatar": {
            "type": "string",
            "example": "https://res.cloudinary.com/dsn4linai/image/upload/v1690095165/fqwwamguzqrsuis7bm1m.png"
          }
        }
      },
...
      "NoticeExtendedResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": "64baea04ba92a3f0060ba5b7"
          },
          "category": {
            "type": "string",
            "enum": ["sell", "lost-found", "for-free", "my-pet"],
            "example": "sell"
          },
          "title": {
            "type": "string",
            "minLength": 4,
            "maxLength": 32
          },
          "age": {
            "type": "string",
            "example": "1 year",
            "description": "Ready calculated age for pet"
          },
          "file": {
            "type": "string",
            "example": "https://res.cloudinary.com/dsn4linai/image/upload/v1690095165/fqwwamguzqrsuis7bm1m.png"
          },
          "sex": {
            "type": "string",
            "enum": ["female", "male"],
            "example": "female"
          },
          "location": {
            "type": "string",
            "minLength": 2,
            "pattern": "^[a-zA-Z]+$",
            "example": "Kyiv"
          },
          "comments": {
            "type": "string",
            "maxLength": 140,
            "default": "",
            "example": "Rich would be the perfect addition to an active family that loves to play and go on walks. I bet he would love having a doggy playmate too! "
          },
          "owner": {
            "$ref": "#/components/responses/UserResponse"
          }
        }
      },
...
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Request body, or params is invalid"
                }
              }
            }
          }
        }
      },
      "NotAuthorized": {
        "description": "Unauthorized user (JWT token missing)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Not authorized"
                }
              }
            }
          }
        }
      },
      "ServerError": {
        "description": "Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Server error"
                }
              }
            }
          }
        }
      }
    }
  }
}